Move repository.api.ts into projectGitLab.api.ts
