+++ client/src/features/project/GitLab.types.ts
@@ -18,6 +18,18 @@
 
 import { Visibilities } from "../../components/visibility/Visibility";
 
+export interface Pagination {
+  currentPage?: number;
+  firstPageLink?: string;
+  lastPageLink?: string;
+  nextPage?: number;
+  nextPageLink?: string;
+  perPage?: number;
+  previousPage?: number;
+  totalItems?: number;
+  totalPages?: number;
+}
+
 export interface GitlabProjectResponse {
   visibility: Visibilities;
   path_with_namespace: string;
@@ -62,4 +74,68 @@ export interface RetryPipelineParams {
 export interface RunPipelineParams {
   projectId: number;
   ref: string;
+}
+
+// GitLab Repository API
+
+export interface GitLabRepositoryBranch {
+  merged: boolean;
+  name: string;
+}
+
+export interface GitLabRepositoryCommit {
+  author_name: string;
+  committed_date: string;
+  id: string;
+  message: string;
+  short_id: string;
+  web_url: string;
+}
+
+export interface GetAllRepositoryBranchesParams {
+  perPage?: number;
+  projectId: string;
+}
+
+export type GetRepositoryBranchResponse = {
+  can_push: boolean;
+  commit: {
+    author_email: string;
+    author_name: string;
+    authored_date: string;
+    committed_date: string;
+    committer_email: string;
+    committer_name: string;
+    created_at: string;
+    id: string;
+    message: string;
+    parent_ids: unknown;
+    short_id: string;
+    title: string;
+    trailers: unknown;
+    web_url: string;
+  };
+  default: boolean;
+  developers_can_merge: boolean;
+  developers_can_push: boolean;
+  merged: boolean;
+  name: string;
+  protected: boolean;
+  web_url: string;
+}[];
+
+export interface GetRepositoryCommitParams {
+  commitSha: string;
+  projectId: string;
+}
+
+export interface GetRepositoryCommitsParams {
+  branch: string;
+  perPage?: number;
+  projectId: string;
+}
+
+export interface GetConfigFromRepositoryParams {
+  commit: string;
+  projectId: number;
 }
