+++ client/src/features/project/projectGitLab.api.ts
@@ -18,19 +18,31 @@
 
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
 import {
+  GetAllRepositoryBranchesParams,
+  GetConfigFromRepositoryParams,
   GetPipelineJobByNameParams,
   GetPipelinesParams,
+  GetRepositoryCommitParams,
+  GetRepositoryCommitsParams,
   GitLabPipeline,
   GitLabPipelineJob,
+  GitLabRepositoryBranch,
+  GitLabRepositoryCommit,
   GitlabProjectResponse,
+  Pagination,
   RetryPipelineParams,
   RunPipelineParams,
 } from "./GitLab.types";
+import processPaginationHeaders from "../../api-client/pagination";
+import { ProjectConfig } from "./Project";
+import { RENKU_CONFIG_FILE_PATH } from "./GitLab.constants";
+import { parseINIString } from "../../utils/helpers/HelperFunctions";
+import { transformGetConfigRawResponse } from "./projectCoreApi";
 
 const projectGitLabApi = createApi({
   reducerPath: "projectGitLab",
   baseQuery: fetchBaseQuery({ baseUrl: "/ui-server/api/projects" }),
-  tagTypes: ["Job", "Pipeline"],
+  tagTypes: ["Branch", "Commit", "Job", "Pipeline"],
   endpoints: (builder) => ({
     // Project API
     getProjectById: builder.query<GitlabProjectResponse, number>({
@@ -40,6 +52,7 @@ const projectGitLabApi = createApi({
         };
       },
     }),
+
     // Project pipelines API
     getPipelineJobByName: builder.query<
       GitLabPipelineJob | null,
@@ -107,6 +120,182 @@ const projectGitLabApi = createApi({
       }),
       invalidatesTags: ["Pipeline"],
     }),
+
+    // Project Repository API
+    getAllRepositoryBranches: builder.query<
+      GitLabRepositoryBranch[],
+      GetAllRepositoryBranchesParams
+    >({
+      queryFn: async (
+        { perPage, projectId },
+        _queryApi,
+        _extraOptions,
+        fetchBaseQuery
+      ) => {
+        const url = `${projectId}/repository/branches`;
+
+        const allBranches: GitLabRepositoryBranch[] = [];
+        let currentPage = 1;
+        // eslint-disable-next-line no-constant-condition
+        while (true) {
+          const result = await fetchBaseQuery({
+            url,
+            params: {
+              page: currentPage,
+              per_page: perPage ?? 100,
+            },
+          });
+
+          if (result.error != null) {
+            return result;
+          }
+
+          const branches = result.data as GitLabRepositoryBranch[];
+          allBranches.push(...branches);
+
+          const responseHeaders = result.meta?.response?.headers;
+          const pagination = processPaginationHeaders(
+            responseHeaders
+          ) as Pagination;
+
+          if (pagination.nextPage == null) {
+            break;
+          }
+
+          ++currentPage;
+        }
+
+        return { data: allBranches };
+      },
+      providesTags: (result) =>
+        result
+          ? [
+              ...result.map(
+                ({ name }) => ({ type: "Branch", id: name } as const)
+              ),
+              "Branch",
+            ]
+          : ["Branch"],
+    }),
+    getConfigFromRepository: builder.query<
+      ProjectConfig,
+      GetConfigFromRepositoryParams
+    >({
+      queryFn: async (
+        { commit, projectId },
+        _queryApi,
+        _extraOptions,
+        fetchBaseQuery
+      ) => {
+        const filePath = encodeURIComponent(RENKU_CONFIG_FILE_PATH);
+        const url = `${projectId}/repository/files/${filePath}/raw?ref=${commit}`;
+        const result = await fetchBaseQuery({ url, responseHandler: "text" });
+
+        if (result.error != null) {
+          return result;
+        }
+
+        const { data, error } = (() => {
+          try {
+            return {
+              data: parseINIString(result.data) as Record<
+                string,
+                Record<string, string>
+              >,
+              error: null,
+            };
+          } catch (error) {
+            return { data: null, error };
+          }
+        })();
+        if (data == null) {
+          return {
+            error: {
+              data: result.data as string,
+              error: `${error}`,
+              originalStatus: result.meta?.response?.status ?? 0,
+              status: "PARSING_ERROR",
+            },
+          };
+        }
+
+        const flattened = Object.entries(data.interactive ?? {}).reduce(
+          (obj, [key, value]) => ({ ...obj, [`interactive.${key}`]: value }),
+          {} as Record<string, string>
+        );
+        const projectConfig = transformGetConfigRawResponse({
+          result: {
+            config: flattened,
+          },
+        });
+
+        return { data: projectConfig };
+      },
+    }),
+    getRepositoryCommit: builder.query<
+      GitLabRepositoryCommit,
+      GetRepositoryCommitParams
+    >({
+      query: ({ commitSha, projectId }) => ({
+        url: `${projectId}/repository/commits/${commitSha}`,
+      }),
+      providesTags: (result) =>
+        result ? [{ id: result.id, type: "Commit" }, "Commit"] : ["Commit"],
+    }),
+    getRepositoryCommits: builder.query<
+      GitLabRepositoryCommit[],
+      GetRepositoryCommitsParams
+    >({
+      queryFn: async (
+        { branch, perPage, projectId },
+        _queryApi,
+        _extraOptions,
+        fetchBaseQuery
+      ) => {
+        const url = `${projectId}/repository/commits`;
+
+        const allCommits: GitLabRepositoryCommit[] = [];
+        let currentPage = 1;
+        // eslint-disable-next-line no-constant-condition
+        while (true) {
+          const result = await fetchBaseQuery({
+            url,
+            params: {
+              ref_name: branch,
+              page: currentPage,
+              per_page: perPage ?? 100,
+            },
+          });
+
+          if (result.error != null) {
+            return result;
+          }
+
+          const commits = result.data as GitLabRepositoryCommit[];
+          allCommits.push(...commits);
+
+          const responseHeaders = result.meta?.response?.headers;
+          const pagination = processPaginationHeaders(
+            responseHeaders
+          ) as Pagination;
+
+          if (pagination.nextPage == null) {
+            break;
+          }
+
+          ++currentPage;
+        }
+
+        return { data: allCommits };
+      },
+      providesTags: (result) =>
+        result
+          ? [
+              ...result.map(({ id }) => ({ type: "Commit", id } as const)),
+              "Commit",
+            ]
+          : ["Commit"],
+    }),
   }),
 });
 
@@ -117,4 +306,8 @@ export const {
   useGetPipelinesQuery,
   useRetryPipelineMutation,
   useRunPipelineMutation,
+  useGetAllRepositoryBranchesQuery,
+  useGetConfigFromRepositoryQuery,
+  useGetRepositoryCommitQuery,
+  useGetRepositoryCommitsQuery,
 } = projectGitLabApi;
