--- client/src/features/repository/repository.types.ts
@@ -1,91 +1,0 @@
-/*!
- * Copyright 2023 - Swiss Data Science Center (SDSC)
- * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
- * Eidgenössische Technische Hochschule Zürich (ETHZ).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export interface Pagination {
-  currentPage?: number;
-  firstPageLink?: string;
-  lastPageLink?: string;
-  nextPage?: number;
-  nextPageLink?: string;
-  perPage?: number;
-  previousPage?: number;
-  totalItems?: number;
-  totalPages?: number;
-}
-
-export interface RepositoryBranch {
-  merged: boolean;
-  name: string;
-}
-
-export interface GetAllRepositoryBranchesParams {
-  perPage?: number;
-  projectId: string;
-}
-
-export type GetRepositoryBranchResponse = {
-  can_push: boolean;
-  commit: {
-    author_email: string;
-    author_name: string;
-    authored_date: string;
-    committed_date: string;
-    committer_email: string;
-    committer_name: string;
-    created_at: string;
-    id: string;
-    message: string;
-    parent_ids: unknown;
-    short_id: string;
-    title: string;
-    trailers: unknown;
-    web_url: string;
-  };
-  default: boolean;
-  developers_can_merge: boolean;
-  developers_can_push: boolean;
-  merged: boolean;
-  name: string;
-  protected: boolean;
-  web_url: string;
-}[];
-
-export interface RepositoryCommit {
-  author_name: string;
-  committed_date: string;
-  id: string;
-  message: string;
-  short_id: string;
-  web_url: string;
-}
-
-export interface GetRepositoryCommitParams {
-  commitSha: string;
-  projectId: string;
-}
-
-export interface GetRepositoryCommitsParams {
-  branch: string;
-  perPage?: number;
-  projectId: string;
-}
-
-export interface GetConfigFromRepositoryParams {
-  commit: string;
-  projectId: number;
-}
