+++ client/src/features/session/hooks/options/useDefaultCommitOption.hook.ts
@@ -19,12 +19,12 @@
 import { useEffect, useMemo } from "react";
 import { useDispatch } from "react-redux";
 import { useLocation } from "react-router";
-import { RepositoryCommit } from "../../../repository/repository.types";
 import { setError } from "../../startSession.slice";
 import { setCommit } from "../../startSessionOptionsSlice";
+import { GitLabRepositoryCommit } from "../../../project/GitLab.types";
 
 interface UseDefaultCommitOptionArgs {
-  commits: RepositoryCommit[] | undefined;
+  commits: GitLabRepositoryCommit[] | undefined;
 }
 
 export default function useDefaultCommitOption({
